Try something with Fraunhofer lines
Not really a good test in hindsight. Here we do not sample a known
distribution but rather really want to bound the TVD from below. While
possible with the squared Hellinger distance it is not favorable the way
we do the estimate for the CDF band. Maybe we just fail to correctly
describe the accuracy of the experiment which has known error bounds
for each interval of the probability function—not a global one.